new feature: new product and refactorings

+++ NossoMercadoLivre/Models/MotherCategory.cs
@@ -2,7 +2,7 @@
 {
     public class MotherCategory
     {
-        public long? CategoryMotherId { get; set; }
+        public int? CategoryMotherId { get; set; }
         public string? Name { get; set; }
     }
 }
